+++ networking/src/lib.rs
@@ -1,12 +1,2 @@
 pub mod rpc;
 pub mod p2p;
-
-mod connection_manager;
-mod network_channel;
-mod peer;
-
-pub mod prelude {
-    pub use super::connection_manager::*;
-    pub use super::network_channel::*;
-    pub use super::peer::*;
-}
