+++ networking/src/p2p/mod.rs
@@ -7,3 +7,5 @@ pub mod peer;
 pub mod pool;
 pub mod client;
 pub mod node;
+pub mod connection_manager;
+pub mod network_channel;
